Restrucre to non-packaged app
