--- app/templates/index.html
@@ -1,13 +1,0 @@
-{% extends "layout.html" %}
-{% block content %}
-<!-- PUT MAP BELOW HERE -->
-
-
-<div>
-  <h1 class="mr-2">Welcome, this is a test header.</h1>
-  <h3 class="mr-2">This is a test subheading</h3>
-</div>
-
-
-<!-- KEEP ALL CODE ABOVE THIS LINE -->
-{% endblock content %}
