--- run.py
@@ -1,6 +1,0 @@
-# Grab our app package
-from app import app
-
-# Debugger
-if __name__ == '__main__':
-    app.run(debug=True)
